+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.1 (unreleased)
 
 - Fixed leading newlines not being removed in one go.
+- Excluded deleted and included untracked files when run on a Git repository.
 
 ## v1.1.0